data/ocr/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt --resolver=backtracking pyproject.toml
#
attrs==22.2.0
    # via pytest
exceptiongroup==1.1.1
    # via pytest
iniconfig==2.0.0
    # via pytest
mss==7.0.1
    # via ocrpii (pyproject.toml)
numpy==1.24.2
    # via opencv-contrib-python
opencv-contrib-python==4.7.0.72
    # via ocrpii (pyproject.toml)
packaging==23.0
    # via
    #   pytesseract
    #   pytest
pillow==9.4.0
    # via
    #   ocrpii (pyproject.toml)
    #   pytesseract
pluggy==1.0.0
    # via pytest
pygame==2.3.0
    # via ocrpii (pyproject.toml)
pytesseract==0.3.10
    # via ocrpii (pyproject.toml)
pytest==7.2.2
    # via ocrpii (pyproject.toml)
tomli==2.0.1
    # via pytest
